Split up huge mail CLI source file
