+++ tools/Nefarius.Utilities.ETW.CLI/NativeConsole.cs
@@ -1,0 +1,16 @@
+namespace Nefarius.Utilities.ETW.CLI;
+
+/// <summary>
+///     Minimal P/Invoke surface for enabling ANSI/VT escape sequences on Windows.
+/// </summary>
+internal static class NativeConsole
+{
+    [System.Runtime.InteropServices.DllImport("kernel32.dll", SetLastError = true)]
+    internal static extern IntPtr GetStdHandle(int nStdHandle);
+
+    [System.Runtime.InteropServices.DllImport("kernel32.dll", SetLastError = true)]
+    internal static extern bool GetConsoleMode(IntPtr hConsoleHandle, out uint lpMode);
+
+    [System.Runtime.InteropServices.DllImport("kernel32.dll", SetLastError = true)]
+    internal static extern bool SetConsoleMode(IntPtr hConsoleHandle, uint dwMode);
+}
